Game CRUD working

+++ parties/parties/src/main/java/com/example/parties/Controllers/GameController.java
@@ -1,7 +1,9 @@
 package com.example.parties.Controllers;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -16,13 +18,16 @@ public class GameController {
     @Autowired
     private GameService gameService;
 
-    @GetMapping("/getGameById")
-    public Game getGameById(@RequestBody Long id) {
+    @GetMapping("/getGameById/{id}")
+    public Game getGameById(@PathVariable Long id) {
         return gameService.findGameById(id);
     }
 
-    @GetMapping("/deleteGame")
-    public void deleteGame(@RequestBody Long id) {
+    @DeleteMapping("/deleteGame/{id}")
+    public void deleteGame(@PathVariable Long id) {
+        if (gameService.findGameById(id) == null) {
+            throw new RuntimeException("Game not found");
+        }
         gameService.deleteGame(id);
     }
